Initial refactoring
Here we create a base node package that defines our app role. A database interface is created, only basic document storage for now. Some handlers are defined to deal with new transactions and new blocks coming in. These are bound to the app.

The spvnode package needs improvement here to allow us to supply a set of handlers in one batch, as opposed to registering each one individually. At the moment it is coupled to the platform/node and this should be decoupled using a handler interface.