Introduce protocol mux
Handlers have been promoted to manage specific protocol actions via the protomux. The listeners bind these handlers to a specific implementation, in this case the spvnode. This could be replaced with anything that puts transactions in and takes transactions out, such as a queue service. A basic verb scheme has been introduced to provide context to the incoming transaction