Integrate wallet in to node
The core node package is now aware of the wallet keys. Some actions may involve two contracts, and these may both be managed by this same instance, so we replay the handler twice for each contract (RootKey) if it is the case. A root key is now passed to every handler, containing the PKH and passed to the responder for signing. The protomux is now fed a promoted inspector transaction, pushing this responsibility up to the orchestrator (spvnode/queue service)